arnold-fitness/arnold-fitness-backend/requirements.txt
boto3
google-generativeai
requests
jsonschema
pydantic
